Create Atm Machine
This program implements the basic functionality of an ATM machine, allowing a user to check their balance, make withdrawals, make deposits, and exit the program. The user is prompted to enter their pin, and if it's correct, they will be presented with a menu of options. The program will continuously loop and display the menu until the user selects the "Quit" option.